fix commit amend
